implement the target for JAXBUtils and createNamespaceWrapper
